Updated to 1.1.2 headers

+++ Source/Starbreeze/Public/SBZEnvQueryGenerator_WorldRuntimeActors.h
@@ -9,10 +9,9 @@ class USBZEnvQueryGenerator_WorldRuntimeActors : public UEnvQueryGenerator {
 public:
 protected:
     UPROPERTY(EditAnywhere, meta=(AllowPrivateAccess=true))
-    uint16 ActorSetting;
+    uint32 ActorSetting;
     
 public:
     USBZEnvQueryGenerator_WorldRuntimeActors();
-
 };
